Ubuntu: Correct product naming to specify release names
Ubuntu releases are named inconsistently in the documentation.
The correct version name is in the format XX.YY where XX is a 2-digit
year and YY is 2-digit month when the product was released.  This is
used in some places but not consistently.  This patch corrects
the cases where only the year is present (e.g. Ubuntu 14 becomes Ubuntu
14.04).  This removed ambiguity as there are 2 releases per year.