Update sysman-service-role.md
The default policy that is suggested here is too broad. It gives, for example, full access to any S3 bucket in the account. In my understanding, everyone with access to the SSM agent server can hack the system to make use of this role and then access anything described in the default policy.
I think we should ask the service team to describe what is the minimum policy for the agent to communicate with the SSM service, and then use this new policy as suggestion.