Bring IAM quickstart's example tags in line with description.
There is a discrepancy between the tags documented (`Key=Finance,Value=WebServers`)
and the keys in the example policy (`Key=tag-key,Value=tag-value`). This change
makes the policy consistent with the paragraph before it.

The divergence happened in 7ded1692 and appears to be at least partially intentional.
I reverted back to the former state for consistency, as there are many other
references to this same set of tags elsewhere in the documentation.