Moved the Important notification
I totally missed that note when reading through the document. So I used the Feedback form to notify the documentation team about that missing. I'm proposing to show it right before a user finds the section on How to proceed with retrieving Secrets using Parameter Store API calls. 
Because I missed that **Important** I made the incorrect assumption that in order for System Manager to process the Secret an ACTUAL Parameter referencing that Secret had to be created first, which is what that note is trying to prevent.