Add a note about maximum username length
When connecting to an instance in Fleet Manager using IAM IC (Formerly AWS SSO) identities, there is a hard maximum of 16 characters for username length - usernames longer will simply be refused with an error message and no workaround. This stems from the Windows AD hard limit of 20, plus AWS prefixing `sso-` to the IAM Identity Center username. This had not previously been indicated in any of the documentation and is important for users to know before making use of this solution.