Correction to minimal policy
The `s3:GetEncryptionConfiguration` is not needed in the minimal policy without SSM session logging (this was confirmed by AWS Support)

The policy is supposed to be minimal and for the scenario where you are not doing logging. Further down the page there is a policy for when you do want logging enabled. Also, I presume the `s3:GetEncryptionConfiguration` action alone would not be enough to allow logging anyway - you'd also need `S3:PutObject` permissions